DL/requirements.txt
shiny
shinywidgets
fastapi
uvicorn[standard]
supabase
httpx
plotly
pandas
python-dotenv
requests
